+++ IsekaiMod/Content/Features/IsekaiProtagonist/SpecialPower/MagicalAmplification.cs
@@ -1,19 +1,15 @@
 ﻿using IsekaiMod.Utilities;
-using Kingmaker.Blueprints.Classes;
 using Kingmaker.Designers.Mechanics.Facts;
-using Kingmaker.Localization;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using TabletopTweaks.Core.Utilities;
 using UnityEngine;
-using static IsekaiMod.Main;
 
 namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower {
 
     internal class MagicalAmplification {
         public static void Add() {
-            LocalizedString MagicalAmplificationDesc = Helpers.CreateString(IsekaiContext, "MagicalAmplification.Description",
-                "Any {g|Encyclopedia:Spell}spells{/g} dealing {g|Encyclopedia:Damage}damage{/g} now change their dice to d10, "
-                + "if it was not d10 or greater. If it already was d10 or greater, instead the spell deals one additional point of damage per dice rolled.");
+            string MagicalAmplificationDesc = "Any {g|Encyclopedia:Spell}spells{/g} dealing {g|Encyclopedia:Damage}damage{/g} now change their dice to d10, "
+                + "if it was not d10 or greater. If it already was d10 or greater, instead the spell deals one additional point of damage per dice rolled.";
             Sprite Icon_TelekineticStrike = BlueprintTools.GetBlueprint<BlueprintAbility>("1a33199538c31a14db23318fdb6e10cb").m_Icon;
 
             var MagicalAmplificationFeature = TTCoreExtensions.CreateToggleBuffFeature(